Print an error and continue execution in a request to the API.
This is needed, because often a request will timeout or something else will happen, and CL ask for a restart. I am not monitoring the scripts 24/7 and sometimes a lot of threads will pile up. In the future, the errors and perhaps a stack trace will may be logged to an error log file.